+++ CHANGELOG.md
@@ -1,2 +1,4 @@
 ## Changelog
+- 11/30/2021 - Add link for recommendation of migration from Protractor
+- 11/30/2021 - Added newrelic-cordova-plugin as a incompatible Capacitor plugin
 - 11/23/2021 - In Capacitor projects: if `cordova-plugin-file-transfer` is used then verify `cordova-plugin-whitelist` is installed (see [https://github.com/ionic-team/capacitor/issues/1199]())